import p5-HTML-Display 0.40 OK benoit@
display HTML locally in a browser

This module abstracts the task of displaying HTML to the user.  The
displaying is done by launching a browser and navigating it to
either a temporary file with the HTML stored in it, or, if possible,
by pushing the HTML directly into the browser window.

Written by: Alexander Bluhm <bluhm@openbsd.org>